requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
selenium>=4.15.0
webdriver-manager>=4.0.0
faker>=20.0.0
factory-boy>=3.3.0
requests>=2.31.0
